Trait darwinia_primitives::IdentifyAccount

pub trait IdentifyAccount {
    type AccountId;
    fn into_account(self) -> Self::AccountId;
}

Some type that is able to be collapsed into an account ID. It is not possible to recreate the original value from the account ID.

Associated Types

type AccountId

The account ID that this can be transformed into.

Loading content...
Required methods

fn into_account(self) -> Self::AccountId

Transform into an account.

Loading content...